Refactor views components
