+++ src/views/NodesView/Nodes/index.ts
@@ -1,0 +1,3 @@
+import DvpnNodes from "./DvpnNodes.vue";
+
+export default DvpnNodes;
