+++ src/views/NodesView/Nodes/tabs/SubscribedNodes.vue
