+++ src/views/SettingsView/SettingsControls/index.ts
@@ -1,0 +1,3 @@
+import SettingsControls from "./SettingsControls.vue";
+
+export default SettingsControls;
