--- src/views/SettingsView/SettingsDns/index.ts
@@ -1,3 +1,0 @@
-import SettingsDns from "./SettingsDns.vue";
-
-export default SettingsDns;
