+++ src/views/SetupCompleteView/SetupComplete/SetupComplete.vue
