+++ src/views/SetupCompleteView/SetupComplete/index.ts
@@ -1,0 +1,3 @@
+import SetupComplete from "./SetupComplete.vue";
+
+export default SetupComplete;
