+++ src/views/SetupCompleteView/SetupCompleteView.vue
@@ -7,6 +7,5 @@
 </template>
 
 <script setup lang="ts">
-import SlrView from "@/components/ui/SlrView/SlrView.vue";
-import SetupComplete from "@/views/SetupCompleteView/SetupComplete.vue";
+import SetupComplete from "@/views/SetupCompleteView/SetupComplete";
 </script>
