--- src/views/SetupGreetingView/types.ts
@@ -1,8 +1,0 @@
-export type View = "greeting" | "setup-complete" | "check-mnemonic";
-
-export type openNextViewMethod = (
-  view: View,
-  config?: {
-    props?: object;
-  }
-) => void;
